+++ AtkCocoa/gailwidget.c
@@ -639,7 +639,7 @@ gail_widget_add_focus_handler (AtkComponent    *component,
                                (gpointer) handler, NULL);
   if (!ret)
     {
-      return g_signal_connect_closure_by_id (component, 
+      return (guint)g_signal_connect_closure_by_id (component,
                                              signal_id, 0,
                                              g_cclosure_new (
                                              G_CALLBACK (handler), NULL,
